Fixed image uploads

+++ backend/src/Utils/cloudinary.js
@@ -14,7 +14,11 @@ const uploadOnCloudinary = async (localFilePath) => {
         const response = await cloudinary.uploader.upload(localFilePath, 
             {resource_type: "auto", transformation: {crop: "scale" }}
         );
-
+        
+        if(!response){
+            console.log("Error uploading file on cloudinary : ", response)
+            return null;
+        }
         // console.log("File is uploaded on cloudinary : ", response.url);
         fs.unlinkSync(localFilePath);
         return response;
